vitral: Use integer coordinates, floats will mess pixel fx

+++ vitral/examples/glium.rs
@@ -35,7 +35,7 @@ mod feature {
         image: vitral::ImageData,
         fore_color: Color,
         back_color: Color,
-        bounds: Rect<f32>,
+        bounds: Rect<i32>,
     }
 
     impl App {
@@ -74,13 +74,8 @@ mod feature {
 
         fn begin(&mut self) { self.core.begin_frame(); }
 
-        fn outline_text(&mut self, pos: Point2D<f32>, text: &str) -> Point2D<f32> {
-            for offset in &[
-                vec2(-1.0, 0.0),
-                vec2(1.0, 0.0),
-                vec2(0.0, -1.0),
-                vec2(0.0, 1.0),
-            ] {
+        fn outline_text(&mut self, pos: Point2D<i32>, text: &str) -> Point2D<i32> {
+            for offset in &[vec2(-1, 0), vec2(1, 0), vec2(0, -1), vec2(0, 1)] {
                 self.core.draw_text(
                     &self.font,
                     pos + *offset,
@@ -94,7 +89,7 @@ mod feature {
                 .draw_text(&self.font, pos, Align::Left, self.fore_color, text)
         }
 
-        fn title_bar(&mut self, bounds: &Rect<f32>, text: &str) {
+        fn title_bar(&mut self, bounds: &Rect<i32>, text: &str) {
             self.core.fill_rect(bounds, self.back_color);
             {
                 let bounds = bounds.inclusivize();
@@ -107,18 +102,18 @@ mod feature {
             }
 
             // Margin
-            let bounds = bounds.inflate(-2.0, -2.0);
+            let bounds = bounds.inflate(-2, -2);
 
             self.core.draw_text(
                 &self.font,
-                bounds.anchor(&point2(0.0, -1.0)),
+                bounds.anchor(&point2(0, -1)),
                 Align::Center,
                 self.fore_color,
                 text,
             );
         }
 
-        fn quit_button(&mut self, bounds: &Rect<f32>) -> bool {
+        fn quit_button(&mut self, bounds: &Rect<i32>) -> bool {
             let click_state = self.core.click_state(bounds);
 
             let color = if click_state != ButtonAction::Inert {
@@ -129,9 +124,9 @@ mod feature {
 
             self.core.fill_rect(bounds, color);
             self.core
-                .fill_rect(&bounds.inflate(-1.0, -1.0), self.back_color);
+                .fill_rect(&bounds.inflate(-1, -1), self.back_color);
 
-            let inner = bounds.inflate(-3.0, -3.0).inclusivize();
+            let inner = bounds.inflate(-3, -3).inclusivize();
 
             self.core
                 .draw_line(1.0, color, inner.bottom_right(), inner.origin);
@@ -144,19 +139,20 @@ mod feature {
 
         fn render(&mut self) {
             self.core
-                .draw_image(&self.image, point2(20.0, 20.0), [1.0, 1.0, 1.0, 1.0]);
-            self.outline_text(point2(22.0, 22.0), "Hello, world!");
+                .draw_image(&self.image, point2(20, 20), [1.0, 1.0, 1.0, 1.0]);
+            self.outline_text(point2(22, 22), "Hello, world!");
         }
     }
 
     pub fn main() {
-        let size = Size2D::new(640.0, 360.0);
+        let size = Size2D::new(640, 360);
         let mut backend: Backend = Backend::start(
             size.width as u32,
             size.height as u32,
             "Vitral Demo",
             vitral::glium_backend::DEFAULT_SHADER,
-        ).expect("Failed to start Glium backend!");
+        )
+        .expect("Failed to start Glium backend!");
 
         let core = vitral::Builder::new().build(size, |img| backend.make_texture(img));
         let mut app = App::new(core, &mut backend);
@@ -165,10 +161,10 @@ mod feature {
             backend.sync_with_atlas_cache(&mut app.atlas_cache);
             app.begin();
             app.render();
-            let (_, title_area) = app.bounds.horizontal_split(12.0);
+            let (_, title_area) = app.bounds.horizontal_split(12);
             app.title_bar(&title_area, "Vitral Demo");
 
-            let (_, widget_area) = title_area.vertical_split(-12.0);
+            let (_, widget_area) = title_area.vertical_split(-12);
             if app.quit_button(&widget_area) {
                 return;
             }
@@ -191,7 +187,8 @@ mod feature {
                         screenshot.width(),
                         screenshot.height(),
                         image::ColorType::RGB(8),
-                    ).unwrap();
+                    )
+                    .unwrap();
                     println!("Screenshot saved!");
                 }
             }
